+++ bc_obps/registration/api/select_operator/user_operator/update_status.py
@@ -1,4 +1,5 @@
-from service.user_operator_service import UserOperatorDataAccessService
+from service.user_operator_service import UserOperatorService
+from registration.api.utils.current_user_utils import get_current_user_guid
 from registration.decorators import authorize, handle_http_errors
 from registration.schema import (
     UserOperatorOut,
@@ -21,4 +22,6 @@ from registration.api.api_base import router
 @authorize(AppRole.get_all_authorized_app_roles(), ["admin"])
 @handle_http_errors()
 def update_user_operator_status(request, payload: UserOperatorStatusUpdate):
-    return 200, UserOperatorDataAccessService.update_user_operator_status(request, payload)
+    return 200, UserOperatorService.update_user_operator_status(
+        payload.user_operator_id, payload, get_current_user_guid(request)
+    )
