+++ bc_obps/registration/api/select_operator/user_operator/user_operator_id.py
@@ -1,5 +1,5 @@
 from uuid import UUID
-from service.user_operator_service import UserOperatorDataAccessService
+from service.data_access_service.user_operator_service import UserOperatorDataAccessService
 
 from registration.decorators import authorize, handle_http_errors
 from registration.schema import (
