+++ bc_obps/registration/api/user_operator/operator.py
@@ -1,6 +1,5 @@
+from service.user_operator_service import UserOperatorService
 from registration.api.utils.current_user_utils import get_current_user_guid
-from service.user_operator_service import UserOperatorDataAccessService
-from service.operator_service import OperatorDataAccessService
 from registration.decorators import authorize, handle_http_errors
 from registration.schema import (
     Message,
@@ -24,7 +23,7 @@ from registration.api.custom_codes_4xx import custom_codes_4xx
 @authorize(["industry_user"], UserOperator.get_all_industry_user_operator_roles(), False)
 @handle_http_errors()
 def create_operator_and_user_operator(request, payload: UserOperatorOperatorIn):
-    return UserOperatorDataAccessService.create_operator_and_user_operator(payload, get_current_user_guid(request))
+    return UserOperatorService.create_operator_and_user_operator(payload, get_current_user_guid(request))
 
 
 ## PUT
@@ -36,6 +35,6 @@ def create_operator_and_user_operator(request, payload: UserOperatorOperatorIn):
 @authorize(["industry_user"], ["admin"])
 @handle_http_errors()
 def update_operator_and_user_operator(request, payload: UserOperatorOperatorIn, user_operator_id: UUID):
-    return 200, UserOperatorDataAccessService.update_operator_and_user_operator(
+    return 200, UserOperatorService.update_operator_and_user_operator(
         user_operator_id, payload, get_current_user_guid(request)
     )
