+++ bc_obps/registration/api/user_operator/operator_access_declined.py
@@ -1,5 +1,5 @@
 from uuid import UUID
-from service.user_service import UserDataAccessService
+from service.data_access_service.user_service import UserDataAccessService
 
 from registration.decorators import authorize, handle_http_errors
 from registration.schema import (
