+++ bc_obps/registration/api/user_operator/user_operator_operator.py
@@ -1,5 +1,5 @@
 from registration.schema.user_operator import PendingUserOperatorOut
-from service.user_service import UserDataAccessService
+from service.data_access_service.user_service import UserDataAccessService
 
 from registration.decorators import authorize, handle_http_errors
 from registration.schema import (
