+++ bc_obps/registration/api/utils/parent_operator_utils.py
@@ -1,5 +1,4 @@
 from typing import List
-from registration.api.utils.operator_utils import handle_operator_addresses
 from registration.models import ParentOperator
 
 
@@ -58,8 +57,9 @@ def handle_parent_operators(updated_parent_operators, operator_instance, user):
             existing_po_mailing_address_id = (
                 existing_po_mailing_address.mailing_address_id if existing_po_mailing_address else None
             )
+            from service.handle_addresses_service import HandleAddressesService
 
-            physical_address, mailing_address = handle_operator_addresses(
+            physical_address, mailing_address = HandleAddressesService.handle_operator_addresses(
                 po_operator.dict(), existing_po_physical_address_id, existing_po_mailing_address_id, 'po_'
             ).values()
