+++ bc_obps/service/application_access_service.py
@@ -1,13 +1,12 @@
 from uuid import UUID
 from registration.schema.user_operator import UserOperatorOut
-from service.user_operator_service import UserOperatorDataAccessService
-from service.user_service import UserDataAccessService
+from service.data_access_service.user_operator_service import UserOperatorDataAccessService
+from service.data_access_service.user_service import UserDataAccessService
 
 
 class ApplicationAccessService:
     def get_operators_business_guid(operator_id: UUID):
         # all approved admins will have the same business_guid so we can use first one
-        breakpoint()
         return UserOperatorDataAccessService.get_approved_admin_users(operator_id).first().business_guid
 
     def does_user_belong_to_operator(operator_id: int, user_guid: UUID):
@@ -21,9 +20,8 @@ class ApplicationAccessService:
         Returns:
             True or raises an exception.
         """
-
         operators_business_bceid = ApplicationAccessService.get_operators_business_guid(operator_id)
-        users_business_bceid = UserDataAccessService.get_user_by_guid(user_guid)
+        users_business_bceid = UserDataAccessService.get_user_by_guid(user_guid).business_guid
 
         if operators_business_bceid != users_business_bceid:
             # brianna this worked great
@@ -43,29 +41,27 @@ class ApplicationAccessService:
             True or raises an exception.
         """
 
-        approved_admins = UserOperatorDataAccessService.get_approved_admins(operator_id)
+        approved_admins = UserOperatorDataAccessService.get_approved_admin_users(operator_id)
 
-        # brianna psuedo
-        if user_guid in approved_admins:
-            raise ("You are already an admin for this Operator!")
+        if approved_admins.filter(user_guid=user_guid).exists():
+            raise Exception("You are already an admin for this Operator!")
 
         if len(approved_admins) > 1:
-            raise ("This Operator already has an admin user!")
+            raise Exception("This Operator already has an admin user!")
 
         # User already has a pending request for this operator
         # NOTE: This is a bit of a weird case, but it's possible for a user to have a pending request for an operator and if we show the UserOperator request form, they could submit another request and end up with two
-        pending_admins = UserOperatorDataAccessService.get_pending_admins(operator_id)
+        pending_admins = UserOperatorDataAccessService.get_pending_admin_users(operator_id)
 
-        # brianna psuedo
-        if user_guid in pending_admins:
-            raise ("You already have a pending request for this Operator!")
+        if pending_admins.filter(user_guid=user_guid).exists():
+            raise Exception("You already have a pending request for this Operator!")
 
         return True
 
     def request_access(operator_id: UUID, user_guid: UUID):
         if ApplicationAccessService.does_user_belong_to_operator(operator_id, user_guid):
             # Making a draft UserOperator instance if one doesn't exist
-            user_operator = UserOperatorDataAccessService.get_or_create_user_operator()
+            user_operator = UserOperatorDataAccessService.get_or_create_user_operator(user_guid, operator_id)
 
         return {"user_operator_id": user_operator.id, "operator_id": user_operator.operator.id}
 
@@ -85,5 +81,5 @@ class ApplicationAccessService:
             # Industry users can only get their own UserOperator instance
             if user_operator.user.user_guid != user_guid:
                 # brianna better message?
-                raise ("Your user is not associated with this operator.")
+                raise Exception("Your user is not associated with this operator.")
             return UserOperatorOut.from_orm(user_operator)
