+++ bc_obps/service/data_access_service/operator_service.py
@@ -8,7 +8,7 @@ class OperatorDataAccessService:
 
     def get_operator_by_user_operator_id(user_operator_id: UUID):
         # brianna why circular import??
-        from service.user_operator_service import UserOperatorDataAccessService
+        from service.data_access_service.user_operator_service import UserOperatorDataAccessService
 
         user_operator = UserOperatorDataAccessService.get_user_operator_by_id(user_operator_id)
         return Operator.objects.get(id=user_operator.operator.id)
